Refactor routes.
