+++ Examples/SampleMinimalApi/API/ToDodoRoutes.cs
@@ -1,0 +1,22 @@
+﻿namespace SampleMinimalApi.API;
+
+public static class ToDodoRoutes
+{
+    public static void UseToDoRoute(this WebApplication app)
+    {
+        var sampleTodos = new Todo[] {
+            new(1, "Walk the dog"),
+            new(2, "Do the dishes", DateOnly.FromDateTime(DateTime.Now)),
+            new(3, "Do the laundry", DateOnly.FromDateTime(DateTime.Now.AddDays(1))),
+            new(4, "Clean the bathroom"),
+            new(5, "Clean the car", DateOnly.FromDateTime(DateTime.Now.AddDays(2)))
+        };
+
+        var todosApi = app.MapGroup("/todos");
+        todosApi.MapGet("/", () => sampleTodos);
+        todosApi.MapGet("/{id}", (int id) =>
+            sampleTodos.FirstOrDefault(a => a.Id == id) is { } todo
+                ? Results.Ok(todo)
+                : Results.NotFound());
+    }
+}
